[项目名称：企小微租户系统][提交人：王晨][目标分解维度返回Map]

+++ qixiaowei-service-api/qixiaowei-system-manage-api/src/main/java/net/qixiaowei/system/manage/api/domain/basic/OfficialRankDecompose.java
@@ -35,7 +35,7 @@ public class OfficialRankDecompose extends BaseEntity {
     /**
      * 具体分解维度的ID
      */
-    private Integer decomposeDimension;
+    private Long decomposeDimension;
     /**
      * 工资系数
      */
